Phase 4 commit 2: extract bootstrap/view-switcher.ts
createViewSwitcher() owns the timeline ↔ notes toggle: the React
notes mount, the visibility flip, the toolbar active-class, and
the history (pushState + popstate + 'notes:exit' event) wiring.
main.ts loses the notesReactRoot module-level state and ~50 lines
of switching logic; the React imports follow.

The DOM scaffold from mountAppShell must be in place before
createViewSwitcher() runs — it queries the shell elements by id.

No behaviour change. Build clean.
